# 1.2.logic optimization