--- a/requirements.txt
+++ b/requirements.txt
@@ -9,6 +9,7 @@ more_itertools
 tatsu>=4.3.0
 hashids>=1.2.0
 jericho>=2.1.1
+mementos>=1.3.1
 
 # For visualization
 pybars3>=0.9.3