--- a/textworld/logic/__init__.py
+++ b/textworld/logic/__init__.py
@@ -19,6 +19,8 @@ from textworld.logic.model import GameLogicModelBuilderSemantics
 from textworld.logic.parser import GameLogicParser
 from textworld.utils import uniquify, unique_product
 
+from mementos import memento_factory, with_metaclass
+
 
 # We use first-order logic to represent the state of the world, and the actions
 # that can be applied to it.  The relevant classes are:
@@ -316,6 +318,7 @@ class TypeHierarchy:
     def __init__(self):
         self._types = {}
         self._children = defaultdict(list)
+        self._cache = {}
 
     def add(self, type: Type):
         if type.name in self._types:
@@ -328,6 +331,9 @@ class TypeHierarchy:
             children = self._children[parent]
             children.append(type.name)
             children.sort()
+
+        # Adding a new type invalidates the cache.
+        self._cache = {}
 
     def get(self, name: str) -> Type:
         return self._types[name]
@@ -435,13 +441,16 @@ class TypeHierarchy:
         """
         return self.multi_closure(types, lambda t: t.child_types)
 
-    def multi_subtypes(self, types: Collection[Type]) -> Iterable[Collection[Type]]:
+    def multi_subtypes(self, types: Collection[Type]) -> List[Collection[Type]]:
         """
         Computes the descendant closure of a sequence of types, including the
         initial types.
         """
-        yield tuple(types)
-        yield from self.multi_descendants(types)
+        types = tuple(types)
+        if types not in self._cache:
+            self._cache[types] = [types] + list(self.multi_descendants(types))
+
+        return self._cache[types]
 
 
 @total_ordering
@@ -522,8 +531,12 @@ class Variable:
         return cls(data["name"], data["type"])
 
 
+SignatureTracker = memento_factory('SignatureTracker',
+                                   lambda cls, args, kwargs: (cls, args[0] + '//'.join(args[1])))
+
+
 @total_ordering
-class Signature:
+class Signature(with_metaclass(SignatureTracker, object)):
     """
     The type signature of a Predicate or Proposition.
     """
@@ -580,8 +593,12 @@ class Signature:
         return _parse_and_convert(expr, rule_name="onlySignature")
 
 
+PropositionTracker = memento_factory('PropositionTracker',
+                                     lambda cls, args, kwargs: (cls, args[0] + '//'.join(v.name for v in args[1]) if len(args) == 2 else ""))
+
+
 @total_ordering
-class Proposition:
+class Proposition(with_metaclass(PropositionTracker, object)):
     """
     An instantiated Predicate, with concrete variables for each placeholder.
     """
@@ -924,7 +941,12 @@ class Action:
         self._pre_set = frozenset(self.preconditions)
         self._post_set = frozenset(self.postconditions)
 
-        self.variables = tuple(uniquify(var for prop in self.all_propositions for var in prop.arguments))
+    @property
+    def variables(self):
+        if not hasattr(self, "_variables"):
+            self._variables = tuple(uniquify(var for prop in self.all_propositions for var in prop.arguments))
+
+        return self._variables
 
     @property
     def all_propositions(self) -> Collection[Proposition]:
@@ -1037,6 +1059,8 @@ class Rule:
         """
 
         self.name = name
+        self.template = None
+        self._cache = {}
         self.preconditions = tuple(preconditions)
         self.postconditions = tuple(postconditions)
 
@@ -1133,9 +1157,19 @@ class Rule:
         The instantiated Action with each Placeholder mapped to the corresponding Variable.
         """
 
+        #key = "//".join(str(mapping[ph]) for ph in self.placeholders)  # Seems faster
+        key = tuple(mapping[ph] for ph in self.placeholders)
+        if key in self._cache:
+            return self._cache[key]
+
         pre_inst = [pred.instantiate(mapping) for pred in self.preconditions]
         post_inst = [pred.instantiate(mapping) for pred in self.postconditions]
-        return Action(self.name, pre_inst, post_inst)
+        action = Action(self.name, pre_inst, post_inst)
+        if self.template:
+            action.template = self.template.format(**{ph.name: "{{{}}}".format(var.name) for ph, var in mapping.items()})
+
+        self._cache[key] = action
+        return action
 
     def match(self, action: Action) -> Optional[Mapping[Placeholder, Variable]]:
         """
@@ -1262,6 +1296,7 @@ class Inform7Logic:
 
     def _initialize(self, logic):
         self._expand_predicates(logic)
+        self._initialize_commands(logic)
 
     def _expand_predicates(self, logic):
         for sig, pred in list(self.predicates.items()):
@@ -1271,6 +1306,14 @@ class Inform7Logic:
                 mapping = {ph: Placeholder(ph.name, type.name) for ph, type in zip(params, descendant)}
                 expanded = pred.predicate.substitute(mapping)
                 self._add_predicate(Inform7Predicate(expanded, pred.source))
+
+    def _initialize_commands(self, logic):
+        for name, command in list(self.commands.items()):
+            rule = logic.rules.get(name)
+            if not rule:
+                continue
+
+            rule.template = command.command
 
 
 class GameLogic: