This PR proposes a few optimizations to speed up the logic engine.

Using the following command
```
python -m cProfile -o before.prof scripts/benchmark_framework.py --game competition/games/test/tw-cooking-recipe5+take5+cook+cut+open+go12-1EnruLgeFPekcB9l.z8
```
**After**
```
Using TWInform7.StateTracking.Inform7Data.GameData.JerichoEnv
Done 1,000 steps in 6.67 secs (150.0 steps/sec) with 12 resets.
Using TWInform7.StateTracking.Inform7Data.GameData.JerichoEnv
Done 1,000 steps in 6.19 secs (161.6 steps/sec) with 7 resets.
Using TWInform7.StateTracking.Inform7Data.GameData.JerichoEnv
Done 1,000 steps in 7.28 secs (137.4 steps/sec) with 10 resets.
-----
Average: 149.7 steps/sec
```
**Before**
```
Using TWInform7.StateTracking.Inform7Data.GameData.JerichoEnv
Done 1,000 steps in 14.21 secs (70.4 steps/sec) with 12 resets.
Using TWInform7.StateTracking.Inform7Data.GameData.JerichoEnv
Done 1,000 steps in 11.94 secs (83.7 steps/sec) with 7 resets.
Using TWInform7.StateTracking.Inform7Data.GameData.JerichoEnv
Done 1,000 steps in 16.37 secs (61.1 steps/sec) with 10 resets.
-----
Average: 71.7 steps/sec
```